refactor transaction to operation
