+++ src/main/java/core/basesyntax/service/OperationProcessor.java
@@ -3,7 +3,7 @@ package core.basesyntax.service;
 import core.basesyntax.model.FruitTransaction;
 import java.util.List;
 
-public interface TransactionProcessor {
+public interface OperationProcessor {
     void process(FruitTransaction transaction);
 
     void processAll(List<FruitTransaction> transactions);
