+++ src/main/java/core/basesyntax/strategy/PurchaseOperationHandler.java
@@ -1,15 +1,14 @@
-package core.basesyntax.service.impl;
+package core.basesyntax.strategy;
 
 import core.basesyntax.db.FruitStorage;
 import core.basesyntax.model.Fruit;
 import core.basesyntax.model.FruitTransaction;
-import core.basesyntax.service.TransactionHandler;
 
-public class PurchaseTransactionHandler implements TransactionHandler {
+public class PurchaseOperationHandler implements OperationHandler {
 
     private FruitStorage storage;
 
-    public PurchaseTransactionHandler(FruitStorage storage) {
+    public PurchaseOperationHandler(FruitStorage storage) {
         this.storage = storage;
     }
