+++ src/main/java/core/basesyntax/strategy/ReturnOperationHandler.java
@@ -1,14 +1,13 @@
-package core.basesyntax.service.impl;
+package core.basesyntax.strategy;
 
 import core.basesyntax.db.FruitStorage;
 import core.basesyntax.model.Fruit;
 import core.basesyntax.model.FruitTransaction;
-import core.basesyntax.service.TransactionHandler;
 
-public class ReturnTransactionHandler implements TransactionHandler {
+public class ReturnOperationHandler implements OperationHandler {
     private FruitStorage storage;
 
-    public ReturnTransactionHandler(FruitStorage storage) {
+    public ReturnOperationHandler(FruitStorage storage) {
         this.storage = storage;
     }
